move test helpers out to separate file
